Add Math Stack Overflow dataset (#42)
This change adds a new dataset for the library judge math resources
server. The new dataset contains math problems from Stack Overflow.

This dataset corresponds to the StackOverflow Dump Data entry in the RL
verifier data tracking
[spreadsheet](https://docs.google.com/spreadsheets/d/1VK4-ZonMSR-4Ulk161Au1f-nGhs4r9V9-beScboaR3I/edit?gid=0#gid=0&range=7:7).

The prompts in the dataset have been formatted in the same way as the
OpenMathReasoning dataset that was previously added. All the Stack
Overflow problems with expected answers were included, so it may be
necessary in the future to filter the problems according to difficulty
based on the responses of a model.

Approval for the use of this dataset is covered in this ticket:
[DGPTT-96](https://jirasw.nvidia.com/browse/DGPTT-96).

Signed-off-by: Damon Mosk-Aoyama <dmoskaoyama@nvidia.com>